XX bumped version, check for file exists before cat'

+++ lib/capistrano/unicorn.rb
@@ -26,30 +26,30 @@ module Capistrano
         desc "stop unicorn"
         task :stop, :roles => :app, :except => { :no_release => true } do
           if pid_running?(unicorn_pid)
-            run "kill `cat #{unicorn_pid}`"
+            run "kill `cat #{unicorn_pid}`" if remote_file_exists(unicorn_pid)
           end
         end
         desc "graceful stop unicorn"
         task :graceful_stop, :roles => :app, :except => { :no_release => true } do
           if pid_running?(unicorn_pid)
-            run "kill -s QUIT `cat #{unicorn_pid}`"
+            run "kill -s QUIT `cat #{unicorn_pid}`" if remote_file_exists(unicorn_pid)
           end
         end
         desc "graceful stop OLD unicorn"
         task :graceful_stop_old, :roles => :app, :except => { :no_release => true } do
           if pid_running?(unicorn_old_pid)
-            run "kill -s QUIT `cat #{unicorn_old_pid}`"
+            run "kill -s QUIT `cat #{unicorn_old_pid}`" if remote_file_exists(unicorn_old_pid)
           end
         end
         desc "reload unicorn"
         task :reload, :roles => :app, :except => { :no_release => true } do
           if pid_running?(unicorn_pid)
             puts "UNICORN RUNNING, reloading"
-            run "kill -s USR2 `cat #{unicorn_pid}`"
+            run "kill -s USR2 `cat #{unicorn_pid}`" if remote_file_exists(unicorn_pid)
             unicorn.graceful_stop_old
           elsif remote_file_exists?(unicorn_pid)
             puts "REMOVING old UNICORN PID"
-            run "rm #{unicorn_pid}"
+            run "rm #{unicorn_pid}" if remote_file_exists(unicorn_pid)
             unicorn.start
           else
             unicorn.start
